Add NPU fused operator
