+++ veomni/ops/npu_patch/hccl_premul_sum.py
